improve vanilla bundle
