+++ package.json
@@ -94,6 +94,7 @@
   "jest": {
     "rootDir": ".",
     "moduleNameMapper": {
+      "^valtio/vanilla$": "<rootDir>/src/vanilla.ts",
       "^valtio$": "<rootDir>/src/index.ts"
     },
     "modulePathIgnorePatterns": [
