+++ tsconfig.json
@@ -8,6 +8,9 @@
     "moduleResolution": "node",
     "baseUrl": ".",
     "paths": {
+      "valtio/vanilla": [
+        "./src/vanilla.ts"
+      ],
       "valtio": [
         "./src/index.ts"
       ]
